requirements.txt
# Backend (Python / pip)
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pandas
python-multipart
python-dotenv
langchain
langchain-groq
bcrypt
python-jose[cryptography]

# Frontend (Node / npm) - install in frontend folder
# npm install react react-dom axios recharts chart.js react-chartjs-2
